requirements.txt
astroid==2.3.2
certifi==2019.9.11
chardet==3.0.4
cycler==0.10.0
Django==2.2.8
django-admin==2.0.0
django-debug-toolbar==2.0
django-excel-response2==2.0.8
django-six==1.0.4
idna==2.8
isort==4.3.21
joblib==0.14.0
kiwisolver==1.1.0
lazy-object-proxy==1.4.2
matplotlib==3.1.1
mccabe==0.6.1
numpy==1.17.2
pylint==2.4.3
pyparsing==2.4.2
python-dateutil==2.8.0
pytz==2019.3
requests==2.22.0
scikit-learn==0.21.3
scipy==1.3.1
screen==1.0.1
six==1.12.0
sklearn==0.0
sqlparse==0.3.0
typed-ast==1.4.0
urllib3==1.25.6
wrapt==1.11.2
xlwt==1.3.0
